Use an integer type for "maxit_" counter.
The "maxit_" counter is an upper limit on the number of non-linear
iterations in a single cell.  Declaring this as a "double" is counter
intuitive unless one expects the number to be *really* high.

Present since
  - Commit 93d4bd8 (TransportModelTwophase.hpp)
  - Commit e0d38cf (TransportModelTwophaseCompressible.hpp)

+++ opm/core/transport/reorder/TransportModelCompressibleTwophase.hpp
@@ -100,7 +100,7 @@ namespace Opm
         std::vector<double> smin_;
         std::vector<double> smax_;
         double tol_;
-        double maxit_;
+        int maxit_;
 
         const double* darcyflux_;   // one flux per grid face
         const double* surfacevol0_; // one per phase per cell
